Refactor dto.creationDto to dto.patent.creationDto
