+++ xml-project/src/main/java/project/xmlproject/dto/patent/creationDto/PodnosilacCreationDto.java
@@ -1,6 +1,5 @@
-package project.xmlproject.dto.creationDto;
+package project.xmlproject.dto.patent.creationDto;
 
-import project.xmlproject.model.patent.KontaktPodaci;
 import project.xmlproject.model.patent.ZahtevZaPriznanjePatenta;
 
 public class PodnosilacCreationDto {
