+++ xml-project/src/main/java/project/xmlproject/service/PatentService.java
@@ -2,7 +2,7 @@ package project.xmlproject.service;
 
 import project.xmlproject.marshal.MarshalPatent;
 import org.springframework.stereotype.Service;
-import project.xmlproject.dto.creationDto.ZahtevZaPriznanjePatentaCreationDto;
+import project.xmlproject.dto.patent.creationDto.ZahtevZaPriznanjePatentaCreationDto;
 import project.xmlproject.model.patent.ZahtevZaPriznanjePatenta;
 import project.xmlproject.repository.PatentRepository;
